[REFACTOR] improve code readability by giving each api module its own name
